requirements.txt
modelscope~=1.23.2
loguru~=0.7.3
langchain-core~=0.3.43
langchain-text-splitters~=0.3.6
llama-index-core~=0.12.23.post2
qdrant-client~=1.13.3
openai~=1.65.4
dotenv~=0.9.9
python-dotenv~=1.0.1
httpx~=0.28.1
fastapi~=0.115.11